LEB128 and Function/Body Split
- Added decoding of signed LEB128
- Using LEB128 for counts, indices, and offsets in several places
- Splitting up function declaration (function header)
  and function body
- Moving body size to head of body per Design PR
- Matching up function bodies with function signatures
- Using UINT32 where getting from LEB128

+++ lib/WasmReader/WasmParseTree.h
@@ -59,7 +59,7 @@ namespace Wasm
 
     struct WasmBlockNode
     {
-        UINT8 count;
+        uint32 count;
     };
 
     struct WasmMemOpNode
@@ -70,14 +70,14 @@ namespace Wasm
 
     struct WasmBrNode
     {
-        UINT8 depth;
+        uint32 depth;
         bool hasSubExpr;
     };
 
     struct WasmTableSwitchNode
     {
-        UINT16 numCases;
-        UINT16 numEntries;
+        uint32 numCases;
+        uint32 numEntries;
         UINT16* jumpTable;
     };
